feat: Separate the guild document from the guild model.
